Move generic Poly algorithms out to src.
